Disable vivante batch processing
With modern kernel drivers, there is no need to implement batching;
this is only necessary if the kernel driver is buggy and allows the
commit-with-stall to return before the pending operations have been
completed.

This allows removal of the batch code; however, keeping the batch
code around is wise should the bug reappear in later code drops.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>