etnaviv: add glyph caching support
Add initial support for glyph caching on a 1024x1024 GPU surface.
This allows us to assemble the glyphs onto a mask buffer for rendering
solely via the GPU rather than using the CPU for this operation.

We always upload the glyphs via a page-aligned page-sized buffer to
avoid any effects from mapping the buffer to the GPU (eg, CPU cache
handling for DMA coherency.)

Signed-off-by: Russell King <rmk@arm.linux.org.uk>