vivante: fix overlapping operations
Where the 2D engine has multiple pixel pipes, it is possible for two
GPU operations to overlap and operate concurrently.  If this occurs
with an overlapping blit copy, one copying to the scanout buffer, and
another copying from the scanout buffer, it results in corrupted pixel
data read from the scanout buffer.

This happens when software cursor mode is enabled with a GC320, as is
the case with iMX6Q SoCs.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>