src: fix failure to enter VT
If EnterVT fails, we end up segfaulting the server, because we have
marked the vtSema true when this isn't the case - we try to LeaveVT
twice, once for the fatal error (which causes a segfault) and once
more as a result of the segfault.  Ensure that we reset this flag
when EnterVT fails to avoid this error.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>