vivante: fix composite using non-alpha pictures
The Vivante GPU hardware appears to be buggy with some formats and
mask combinations, and is particularly noticable with BGRA/RGBA
swizzles with non-alpha formats.  Fix this by switching to alpha
formats, and subsituting a maximal global alpha value instead.  This
allows rendercheck to pass for all formats with PE20 hardware.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>