etnaviv: add component alpha blending for PE2.0 hardware
PE2.0 hardware has support for component alpha blending.  This is where
the blend factors are taken from each individual channel of the other
image, rather than the alpha channel of the other image.  Hence, a
standard InReverse would be:

	dst.A = src.A * 0 + dst.A * src.A
	dst.C = src.C * 0 + dst.C * src.A

A component alpha InReverse is:

	dst.A = src.A * 0 + dst.A * src.A
	dst.C = src.C * 0 + dst.C * src.C

This mode is selected by the COLOR and COLOR_INVERSED blend modes.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>