etnaviv: avoid PutImage() bringing a pixmap onto GPU
If PutImage() is called on a pixmap which is not already on the GPU,
do not pull it onto the GPU.  Delay this until the pixmap is used
some other way.  This avoids unnecessarily bouncing a pixmap onto
the GPU at PutImage time only to have it freed a short time later.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>