etnaviv: add etnaviv DRM support
Add support for Christian's Etnaviv DRM interface.  We add this as a
shim between the conventional etnaviv APIs and the kernel DRM
interfaces rather than using libdrm-etnaviv.

The reasoning here is that libdrm-etnaviv requires changes to the
conventional etnaviv interfaces, which make it difficult to build a
common driver core and share that between a pure etnaviv and a libdrm
implementation.

Hence, this includes Christian's etnaviv_drm.h header, with a few
modifications to allow it to be included along side conventional
etnaviv, and also adding support for userptr mapping.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>