etnaviv: temporarily disable tiling for glyph cache
Tests on GC320 reveal that it seems not to support tiling for the glyph
cache.  Having tiling enabled means that we quadruple the pitch, which,
if the GPU does not support tiling, has the side effect of causing the
GPU to read/write beyond the bounds of the pixmap, corrupting memory.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>