etnaviv: provide etnaviv_compute_composite_region()
miComputeCompositeRegion() needs the coordinates transformed to the
backing pixmap, and then pre-translated via the transformation.  Since
miComputeCompositeRegion() can only cope with a translation transform,
we arrange for etnaviv_compute_composite_region() to return -1 should
there be a non-translation transform, or the true/false value from
miComputeCompositeRegion().

In doing this, we need to reposition our call to
miComputeCompositeRegion() and tweak things a little to ensure that
the coordinates we're using are properly adjusted at the appropriate
points in the code.

Since we reject non-translation transforms early, we can eliminate the
now duplicated test in the mask processing code.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>