etnaviv: fallback for etnaviv_accel_composite_masked()
We can gain some performance for cases we don't handle by performing
the masked blend in software, but still having the final blend in
hardware.

Rearrange etnaviv_accel_composite_masked() so that we are able to setup
what we need, and then for cases we don't handle, fall back to using
a software blend instead.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>