etnaviv: remove transformations at clip time
Extensive analysis of Xorg, Intel SNA, and other code indicates that
the source clips are not affected by the transformation matrix, only
the drawable position on the backing pixmap needs to be included in
the passed coordinates.

Since we now let through all transformations, we must ensure that
transformations we can't handle in hardware cause us to fallback.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>