etnaviv: etnadrm: add support for GEM_WAIT ioctl
Add support for the new GEM_WAIT ioctl, which waits for the kernel to
finish with the GEM object (all render complete, and has been retired).
This is different from the WAIT_FENCE ioctl, which just waits for
rendering to complete.

This is an important distinction, as when a gem object is retired and
freed, the kernel will invalidate the cache for this object, which
effectively changes the data userspace sees.  If userspace merely
waits for rendering to complete, there is a race condition where a
userptr BO can be free()d and re-used, and the kernel's invalidation
then corrupts the malloc() free lists.

This commit ensures that etna_bo_del() on a userptr BO will not
return until the kernel has done with the BO.

On kernels which do not support the GEM_WAIT ioctl, we presently
ignore the error.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>