etnaviv: delay freeing the userptr etna bo and memory
Delay freeing the userptr etna bo and memory until we're about to
block, rather than waiting for the glyph upload to complete.  This
should make glyph upload of multiple glyphs more efficient as we
don't have to wait for each to complete before uploading the next
one.

The down-side is an increase in the memory used by the X server, as
we have to keep this memory around until the block handler gets
called.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>