etnaviv: convert etnaviv_blit_srcdst() to use absolute source origin mode
etnaviv_blit_srcdst() is a single-box copy from a source origin point
to a single box on the destination.  It doesn't make sense to calculate
a relative offset for the source, we might as well use absolute origin
mode.  Switch to absolute origin mode, and add a DE op helper for this
operation.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>