src: Xv: allow Xvbo name import to be overridden
If we have a DRM-based GPU, the DRI2 system exports the GPU DRM device
rather than the KMS DRM device, and so the Xvbo names will be specific
to the GPU rather than the KMS.  Therefore, we need a way to talk to
our accelerator backend to convert the GPU name into something we can
display.

This adds a hook in the Xv code to allow this, updates the GPU drivers
to return capabilities indicating what they support.  Currently, the
etnaviv/drm backend does not support Xvbo in any form.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>