etnaviv: Xv: Add 1 to XV_ENCODING
When mplayer starts up it queries all properties and divides the min by
the max.  With XV_ENCODING having a range of 0 - 0, this causes a
divide-by-zero error in mplayer.

Since many people reach for mplayer when they want to play video, hack
the XV_ENCODING property to have a nonzero difference by making the max
report as being 1.

Signed-off-by: Sean Cross <xobs@kosagi.com>