etnaviv: add DRI3 support
Add DRI3 support to the etnadrm backend, defaulting to be disabled as
this has no capability to support flipping the scanout pixmap on split
KMS/render hardware.  DRI2 is preferred for this.

A new option is added to allow DRI3 support to be enabled if DRI3
support is desired, in a manner similar to other DDX.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>