src: add support for Xorg platform bus and non-root servers
Add support for the Xorg platform bus DRM device discovery, which
allows DDX to work on systems that run the X server as a non-root user.
This adapts the KMS device only to this new model.

This is made more complex by bugs in the X server: the X server
proceeds to call the legacy probe method even after screens have been
discovered via the platformProbe method when the last device to be
probed fails (eg, due to the last device being a GPU device.)

We work around that by always opening the DRM device at probe time, and
issuing a SetVersion call, which causes subsequent probe attempts to
fail.

If we are running on a server which passes us the DRM device's file
descriptor, or a server running non-root, we must not attempt to drop
and regain the DRM master status ourselves; this will be managed for
us by the X server and other system components.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>