etnaviv: more efficient state buffer building
Analysis shows that we can gain some additional performance by building
the GPU state buffer more efficiently - most of the overhead comes from
the assert() calls on each emission to the state buffer.  Avoid this by
pre-checking and post-checking the state size, and doing as few of
these checks as possible.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>