etnaviv: fix size calculation in EL_START()
Dennis Gilmore reports that Fedora sees assert()s from the etnaviv
backend when XFCE starts up.  This is due to an incorrect calculation
in EL_START() which mistakenly multiplied the operation size by 4
(sizeof(uint32_t)) before adding it to the current batch buffer index.
This resulted in the calculated size overflowing the maximum batch
size, and causing the false assert().

Signed-off-by: Russell King <rmk@arm.linux.org.uk>